refactor: move FullRpcProvider trait to storage-api crate
