+++ crates/storage/storage-api/src/lib.rs
@@ -102,3 +102,6 @@ pub use state_writer::*;
 
 mod header_sync_gap;
 pub use header_sync_gap::HeaderSyncGapProvider;
+
+mod full;
+pub use full::*;
